idint

+++ finalWebbApp/server/api/src/main/java/com/daoAPI/IBaseDAO.java
@@ -4,7 +4,7 @@ package com.daoAPI;
 public interface IBaseDAO <T> {
     void addEntity(T entity) throws Exception;
     void updateEntity(T entity) throws Exception;
-    void deleteEntity(Long id) throws Exception;
+    void deleteEntity(Integer id) throws Exception;
     // List<T> getAllEntities( TypeSorting sorting);
-     T getEntityById(Long id) throws Exception;
+     T getEntityById(Integer id) throws Exception;
 }
