+++ finalWebbApp/server/api/src/main/java/com/daoAPI/IUserDAO.java
@@ -5,7 +5,7 @@ import com.model.User;
 import org.hibernate.Session;
 
 public interface IUserDAO extends IBaseDAO<User> {
-    Long checkUser(String login)throws Exception;
+    Integer checkUser(String login)throws Exception;
     User getUserByLoginPassword(String login,String password)throws Exception;
     void removeUser(User user) throws Exception;
     Session getSession();
